+++ doc/source/changelog.rst
@@ -18,6 +18,7 @@ master: 1.7.1
 * Fix by-ptr return of objects in Python derived classes
 * Fix for passing numpy boolean array through ``bool*``
 * Fix assignment to ``const char*`` data members
+* Allow passing sequence of strings through ``const char*[]`` argument
 
 
 2020-04-27: 1.7.0